Update ggit_blob_get_raw_size() to return a gsize
libgit2 used to return an int now it returns a size_t.

+++ libgit2-glib/ggit-blob.h
@@ -60,7 +60,7 @@ GgitBlob        *_ggit_blob_wrap             (git_blob *blob,
 
 gconstpointer     ggit_blob_get_raw_content  (GgitBlob *blob);
 
-gint              ggit_blob_get_raw_size     (GgitBlob *blob);
+gsize             ggit_blob_get_raw_size     (GgitBlob *blob);
 
 G_END_DECLS
